Rename file names
Bench: 2266395
